Feat : searchView 검색어 연동, postScreen 임의로 입력 삭제, 메시지 추가, 실시간 촬영 checkbox 변경
