+++ presentation/src/main/java/com/pat/presentation/ui/pat/PatDetailScreen.kt
@@ -250,7 +250,6 @@ fun PostDetailScreen(
                 textColor = White,
                 onClick = {
                     patDetailViewModel.participatePat()
-                    navigation()
                 }
             )
         }
